Nurse expected days/month: decimal default 23.83 (DR 44h legal working days)
HomeCareMonthlyExpectedDays int->decimal (default 23.83 = 5.5 days/week x 52 / 12, the Código de Trabajo divisor for daily/hourly pay). Daily = HomeCareMonthlyRate / HomeCareMonthlyExpectedDays is now legally exact. Updated domain, 4 DTOs ([Range(1.0,31.0)]), both ComputeNurseBasePay overloads + NurseRates, seeding, and the profile-completion param. Migration AlterNurseExpectedDaysDecimal applied (column decimal(18,2); existing 30 -> 30.00, no data loss). Build clean; API round-trips decimal.

+++ src/NursingCareBackend.Application/Identity/Responses/NurseProfileAdminResponse.cs
@@ -26,6 +26,6 @@ public sealed record NurseProfileAdminResponse(
     string? Category,
     decimal VisitDailyRate,
     decimal HomeCareMonthlyRate,
-    int HomeCareMonthlyExpectedDays,
+    decimal HomeCareMonthlyExpectedDays,
     NurseWorkloadSummary Workload
 );
